Create the_code.dart
In this example:

The fetchData function is an asynchronous function that uses the http package to make a GET request to the specified API endpoint (apiUrl). It then parses the JSON response and returns a Map<String, dynamic>.

The FutureBuilder widget is used to asynchronously build a widget tree based on a Future.

While the data is being fetched (ConnectionState.waiting), a loading indicator is displayed.

If an error occurs during the fetch, an error message is displayed.

If the fetch is successful, the retrieved data is displayed.